# Prevent multiple enumeration in GenericArgsSatisfy